+++ 03_Parsing/parser_rd_helper_b.c
@@ -61,3 +61,58 @@ t_ast_node	*create_redir(t_parser_context *context)
 		node = create_ast_node(AST_REDIR_APPEND);
 	return (node);
 }
+
+/**
+ * @function: handle_redirection_target
+ * @brief: Creates and processes the target node for redirection
+ * 
+ * @param context: The parser context containing token information
+ * @param node: The parent redirection node
+ * @return: void
+ */
+void	handle_redirection_target(t_parser_context *context, t_ast_node *node)
+{
+	t_ast_node	*target_node;
+
+	if (is_token_type(context, TOKEN_RD_FD))
+	{
+		target_node = create_ast_node(AST_RD_FD);
+		target_node->value = ft_strdup(((t_lex_data *)
+					context->current_token->content)->raw_string);
+		add_child_node(node, target_node);
+		advance_token(context);
+	}
+	else
+	{
+		context->error = 1;
+		context->error_message = ft_strdup("Expected a redirection target");
+	}
+}
+
+/**
+ * @function: handle_heredoc_delimiter
+ * @brief: Creates and processes the delimiter node for heredoc
+ * 
+ * @param context: The parser context containing token information
+ * @param node: The parent heredoc node
+ * @return: void
+ */
+void	handle_heredoc_delimiter(t_parser_context *context, t_ast_node *node)
+{
+	t_ast_node	*target_node;
+
+	if (is_token_type(context, TOKEN_HD_DELIMITER_Q))
+		target_node = create_ast_node(AST_HD_DELIMITER_Q);
+	else if (is_token_type(context, TOKEN_HD_DELIMITER_NQ))
+		target_node = create_ast_node(AST_HD_DELIMITER_NQ);
+	else
+	{
+		context->error = 1;
+		context->error_message = ft_strdup("Expected a heredoc delimiter");
+		return ;
+	}
+	target_node->value = ft_strdup(((t_lex_data *)
+				context->current_token->content)->raw_string);
+	add_child_node(node, target_node);
+	advance_token(context);
+}
